Adicionando cache
