+++ src/main/java/com/fiap/parquimetro/api/resource/condutor/CondutorResource.java
@@ -1,10 +1,11 @@
 package com.fiap.parquimetro.api.resource.condutor;
 
-import com.fiap.parquimetro.dominio.gerencimantoCadastro.condutor.dto.DadosCondutor;
-import com.fiap.parquimetro.dominio.gerencimantoCadastro.condutor.entity.Condutor;
-import com.fiap.parquimetro.dominio.gerencimantoCadastro.condutor.service.CondutorService;
+import com.fiap.parquimetro.dominio.gerencimentoCadastro.condutor.dto.DadosCondutor;
+import com.fiap.parquimetro.dominio.gerencimentoCadastro.condutor.entity.Condutor;
+import com.fiap.parquimetro.dominio.gerencimentoCadastro.condutor.service.CondutorService;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
+import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +18,6 @@ public class CondutorResource {
 
     private final CondutorService condutorService;
 
-
     @PostMapping()
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<DadosCondutor> cadastrar(@Valid @RequestBody DadosCondutor dadosCondutor,
@@ -28,5 +28,13 @@ public class CondutorResource {
 
         return ResponseEntity.created(uri).body(new DadosCondutor(condutor));
 
+    }
+
+    @GetMapping("/{cpf}")
+    public ResponseEntity<DadosCondutor> buscarPorCpf(@PathVariable String cpf) {
+        return  condutorService.carregar(cpf)
+                .map(DadosCondutor::new)
+                .map(ResponseEntity::ok)
+                .orElse(ResponseEntity.notFound().build());
     }
 }
