--- src/main/java/com/fiap/parquimetro/dominio/gerencimantoCadastro/condutor/repository/CondutorRepository.java
@@ -1,7 +1,0 @@
-package com.fiap.parquimetro.dominio.gerencimantoCadastro.condutor.repository;
-
-import com.fiap.parquimetro.dominio.gerencimantoCadastro.condutor.entity.Condutor;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CondutorRepository extends JpaRepository<Condutor, Long> {
-}
