--- src/main/java/com/fiap/parquimetro/dominio/gerencimantoCadastro/veiculo/dto/DadosVeiculo.java
@@ -1,33 +1,0 @@
-package com.fiap.parquimetro.dominio.gerencimantoCadastro.veiculo.dto;
-
-import com.fiap.parquimetro.dominio.gerencimantoCadastro.condutor.entity.Condutor;
-import com.fiap.parquimetro.dominio.gerencimantoCadastro.veiculo.entity.Veiculo;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-
-public record DadosVeiculo(
-
-        Long id,
-
-        @NotBlank
-        String placa,
-
-        @NotBlank
-        String fabricante,
-
-        @NotBlank
-        String modelo,
-
-        @NotBlank
-        String cor,
-
-        @NotEmpty
-        @NotNull
-        Condutor Condutor
-
-) {
-    public DadosVeiculo(Veiculo veiculo) {
-        this(veiculo.getId(), veiculo.getPlaca(), veiculo.getFabricante(), veiculo.getModelo(), veiculo.getCor(), veiculo.getCondutor());
-    }
-}
