--- src/main/java/com/fiap/parquimetro/dominio/gerencimantoCadastro/veiculo/repository/VeiculoRepository.java
@@ -1,7 +1,0 @@
-package com.fiap.parquimetro.dominio.gerencimantoCadastro.veiculo.repository;
-
-import com.fiap.parquimetro.dominio.gerencimantoCadastro.veiculo.entity.Veiculo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
-}
