+++ src/main/java/com/fiap/parquimetro/dominio/gerencimentoCadastro/cartao/repository/CartaoRepository.java
@@ -1,0 +1,12 @@
+package com.fiap.parquimetro.dominio.gerencimentoCadastro.cartao.repository;
+
+import com.fiap.parquimetro.dominio.gerencimentoCadastro.cartao.entity.Cartao;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface CartaoRepository extends JpaRepository<Cartao, Long> {
+
+    List<Cartao> findCartaoByCondutorId(Long condutorId);
+
+}
