+++ src/main/java/com/fiap/parquimetro/dominio/gerencimentoCadastro/condutor/enumeration/FormaPagamento.java
@@ -1,0 +1,15 @@
+package com.fiap.parquimetro.dominio.gerencimentoCadastro.condutor.enumeration;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum FormaPagamento {
+
+    CREDITO("Crédito"),
+    DEBITO("Débito"),
+    PIX("PIX");
+
+    private String nome;
+}
