+++ src/main/java/com/fiap/parquimetro/dominio/gerencimentoCadastro/veiculo/repository/VeiculoRepository.java
@@ -1,0 +1,7 @@
+package com.fiap.parquimetro.dominio.gerencimentoCadastro.veiculo.repository;
+
+import com.fiap.parquimetro.dominio.gerencimentoCadastro.veiculo.entity.Veiculo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
+}
