+++ src/main/resources/application.properties
@@ -1,17 +1,13 @@
 spring.application.name=parquimetro
 
-#spring.datasource.url=jdbc:h2:mem:parquimetro
-#spring.datasource.username=sa
-#spring.datasource.password=
-#spring.h2.console.enabled=true
-#spring.h2.console.path=/h2-console
-#
 #spring.jpa.hibernate.ddl-auto=update
 #
-#spring.jpa.show-sql=false
-#spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
 
 spring.datasource.url=jdbc:postgresql://localhost:5432/parquimetro
 spring.datasource.username=${USER_DB}
 spring.datasource.password=${PASSWORD_DB}
 spring.jpa.hibernate.ddl-auto=update
+
+spring.cache.cache-names=CondutorCache, VeiculoCache, EstacionamentoCache
