+++ src/main/resources/messages.properties
@@ -11,3 +11,4 @@ Email=deve ser um e-mail válido
 dataIntegrityViolationMessage=Ocorreu um erro ao processar sua solicitação. Parece que os dados fornecidos violam uma restrição de integridade. Por favor, verifique os dados e tente novamente.
 constraintViolationMessage=Ocorreu um erro de validação. Por favor, verifique os dados e tente novamente.
 methodArgumentNotValidMessage=Ocorreu um erro de validação. Por favor, verifique os dados e tente novamente.
+httpRequestMethodNotSupportedException=O método não é suportado para este endpoint.
